requirements.txt
Flask==2.3.1
MarkupSafe==2.1.2
Flask-Session==0.8.0
faiss-gpu==1.7.2
groq
requests
azure-cosmos==4.7.0
fastapi
DateTime
python-dotenv
azure-functions

